images (icons and fonts) are now inside the binary.
Added the application dir2hcpp which creates a .h and .cpp file
from a directory. This is used to generate header and source files
containing the data of the images and fonts.
Therefore bpv does not require the directories images and fonts during
runtime anymore.
The generated files will be stored in include_gen and source_gen.
To compile the code from scratch, just do "qmake bpv.pro" followed
by "make release" (or "mingw32-make release")