Fix BPV for Qt5, add README.md, fix bug with .geo
QString doesn't have the toAscii() function in Qt5.
Add a README.md for github.
Fix a bug with geometries given in .geo files:
when using the 'transform' option with a matrix
having a negative determinent, the triangles
generated in OpenGL would created in the wrong order
making the primitive black. Reorder the vertices
according to the sign of the determinent.